First uploaded version
It's all built in illustrator, so the elements are all separate vectors that can be rearranged to your heart's content. I am not good at actual hand-drawing so I faked the handmade look with brushes & offsetting edges of objects. It's as tidy as I could make it while still keeping the aesthetic a bit messy.